+++ include/update/updater.h
@@ -25,7 +25,6 @@
 
 #include <mutex>
 #include <string>
-#include "network/webclient.h"
 #include "version.h"
 
 namespace Nickvision::Update
@@ -83,7 +82,6 @@ namespace Nickvision::Update
 
     private:
         mutable std::mutex m_mutex;
-        Network::WebClient m_webClient;
         std::string m_repoOwner;
         std::string m_repoName;
         int m_latestStableReleaseId;