+++ src/network/web.cpp
@@ -1,33 +1,10 @@
-#include "network/webclient.h"
+#include "network/web.h"
 #include <fstream>
 #include <sstream>
 
 namespace Nickvision::Network
 {
-    bool WebClient::m_initialized{ false };
-    unsigned int WebClient::m_instanceCount{ 0 };
-
-    WebClient::WebClient()
-    {
-        if(!m_initialized)
-        {
-            curl_global_init(CURL_GLOBAL_ALL);
-            m_initialized = true;
-        }
-        m_instanceCount++;
-    }
-
-    WebClient::~WebClient()
-    {
-        m_instanceCount--;
-        if(m_initialized && m_instanceCount == 0)
-        {
-            curl_global_cleanup();
-            m_initialized = false;
-        }
-    }
-
-    bool WebClient::getWebsiteExists(const std::string& url) const
+    bool Web::getWebsiteExists(const std::string& url)
     {
         if(url.empty())
         {
@@ -38,7 +15,7 @@ namespace Nickvision::Network
         return curl.perform() == CURLE_OK;
     }
 
-    std::string WebClient::fetchJson(const std::string& url) const
+    Json::Value Web::fetchJson(const std::string& url)
     {
         if(url.empty())
         {
@@ -49,10 +26,23 @@ namespace Nickvision::Network
         curl.setUserAgent("Mozilla/5.0 (Windows NT 10.0; Win64; x64; rv:109.0) Gecko/20100101 Firefox/120.0");
         curl.setHeaders({ "Content-Type: application/json" });
         curl.setStream(&out);
-        return curl.perform() == CURLE_OK ? out.str() : "";
+        if(curl.perform() == CURLE_OK)
+        {
+            std::string data{ out.str() };
+            if(!data.empty())
+            {
+                Json::Value root;
+                Json::Reader reader;
+                if(reader.parse(out.str(), root, false))
+                {
+                    return root;
+                }
+            }
+        }
+        return {};
     }
 
-    bool WebClient::downloadFile(const std::string& url, const std::filesystem::path& path, const CurlProgressFunction& progress, bool overwrite) const
+    bool Web::downloadFile(const std::string& url, const std::filesystem::path& path, const CurlProgressFunction& progress, bool overwrite)
     {
         if(url.empty())
         {